[search] Improve search ranking behavior
This PR fixes some bugs in search rankings, and should make search more
usable, especially for partial matches and buildings whose numbers aren't
yet in OSM. Here's a summary of changes:

Search can now recognize rt=route, ct=court, and a few other synonyms

The current version of search pops the last word of a query and saves it
to a variable called m_prefix. The prefix token was not considered when
ranking results, so a search for "S Valley Way" would be functionally
truncated to "S Valley". As a result, the search was unable to differentiate
between Way/St/Circle/etc.

Searching for an address as "1600 Pennsylvania Ave" is very common, and the
prefix handling made that very inaccurate.

That said... I'm American. I'm not sure if the prefix handling serves an
important role for users of other nationalities. Let me know if so and I'll
revisit it. (All code to handle prefixes is still there; I just disabled it
in `search/processor.cpp`)

The current search computes matchLength based on the name of matched feature.
That meant a search for "S Xenia St" could match "South Valley Highway" - they
share the "South". But the highway is 18 characters long. That produced a
matchedFraction of 18/8, which incorrectly increased the relevance of that match.

The new implementation computes and stores matchedLength as part of the
GetNameScores routine. It computes match length on a token-by-token basis
and uses the query length instead of feature length. In the prior example,
that means that only the letter S matches. That yields a 1/8 matchedFraction score,
which is much more reasonable.

Migrating matchedLength into nameScores allows `GetNameScores` to return just a
`NameScores` instead of a `pair<NameScores, size_t>`. That simplified accessing the data.

The change also allowed for removing a few other places that computed matchLength,
such as `ranker.cpp:195`.

The current implementation removes typing suggestions from search results.
While that prevents duplication, it also keeps users from accessing the highly-ranked
result. They'd have to click the suggestion and wait for the search to rerun.

The new implementation leaves suggestions alone, but also presents the
suggestions as normal results.

Most of the ranking logic lies in the ranking_utils.hpp implementation of
`GetNameScores`. I was struggling to understand it as-was due to nondescript
variable names and missing comments. The new version is heavily commented
and behaves very similarly in unit testing.

There are a few new test cases to explicitly test synonyms (s/south, st/street, etc).
All tests now check matchedLength in addition to error count. One was removed:
```
   test("San Francisco", "Fran ", TokenRange(0, 1), NAME_SCORE_ZERO, 0);
```
failed since the tokenizing process strips spaces. Once spaces are removed,
"Fran" does match "San Francisco".

Let me know what you think! Between playing around in unit testing and
the desktop app, I'm much happier with the search results I'm seeing.

Signed-off-by: Ben Nitkin <ben@nitkin.net>